refactor the manager.
Now the manager only take the loop to use on initialization and should
be started using the `start`. Controllers are now passed to the start
function as well. This will the use of the Manager with any lib using
pyuv. While I'm here remove the rpc feature, it can be setup using a
controller.