add IO redirection
It is now possible to redirect IO (stdout & stderr) to an event feed.

The IO rediction are setup using the stdio parameter in the process
template:

    m = Manager()
    m.add_process("pname", "/some/cmd", stdio=["a", "b"])

Here ["a", "b"]Â will redirect stdoutt & stderr and stdout events will
be labeled "a". Any label can be used the maximum number of redirect io
are 2, stderr can be ignored. If you want to redirect stderr to stdout
just use the same label.

Then to subscribe to an io event:

    p.monitor_io("a", listener_cb)

p is one running process using the "name" template. listener_cb is a
callable that will get (event, msg)

for ex:

    ('a', {'data': b'hello err', 'pid': 1, 'name': 'pname', 'event':
'a'})