add webhooks support
Add a new gaffer application to handle webhooks. Each events can triger
a post on a given url.

for example to listen all create events on http://echohttp.com/echo you
can add this line in the webhooks sections of the gaffer setting file:

    [webhooks]
    create = http://echohttp.com/echo you

Or programatically:

    from gaffer.manager import Manager
    from gaffer.webhooks import WebHooks
    hooks = [("create", "http://echohttp.com/echo you ")
    webhooks = WebHooks(hooks=hooks)

    manager = Manager()
    manager.start(apps=[webhooks])

All events are supported.