handle groups of processes
This patch add a the possibility to manipulate group of processes in the
manager. 1 process can be attached to 1 group only. Only one process
with the same name can be in a group.

To keep the process template name unique groups are defined in it under
the format `<group>:<name>` .