make `Process.stats more accurate.
start to cycle the cpu stats so we can have an accurate number on the first
call of ``Process.stats``. Since this is a blocking IO call, do it in a
thread.

Note: there is still a possibility to call it before the queued thread have
been released, but it should be rare. On the worst case 0.0 will be returned.