`gaffer unload` replace gafferctl command like.
gaffer unload can unload all jobs from a procfile or from a JSON config file.
Also all jobs name passed on the command line can be unloaded. In this cases
jobs can be either defined in the procfile or be loaded in a gafferd node.

While I'm here fix the load command to reflect changes in the unload commad:
- now if a job already exists fro this app it prints a conflict.
- you can now pass multiple jobs in a josn config file, see the file
  `jobs.json` in examples.