simplify the HTTP handling.
There is no real need to listen on multiple endpoints here. It can easily be
done either by using masquerading rules or a prox. This change make the code
more simple and prepare the arrivale of the lookupd server.

While I'm here make sure an unix adress can't be used by a gaffer node since
we won't be abble to register it in a lookupd server anyway.