simplify the event emitter
This change simplify the event listener code by sending wildcards and patterns
subscribed all at once. There is no real reason to send them in a different
loop also the way we trigger thevent loop using an async worker may be a
problem when both a wildcard and a simple event were subcscribed.