make sure we trigger all events.
following @saghul idea, start the spinner, pass it a dummy callback (lambda h:
None) and stop it at the end of _send. This will prevent the loop from
blocking for i/o on a given iteration if there are pending events, that is,
until the prepare handle is run. Als unref the prepare handle to not block the
loop.